Text controller, repository and service fo formality service.

+++ formality-service/src/main/java/pl/tul/emailsemotions/formalityservice/repositories/TextRepository.java
@@ -3,6 +3,6 @@ package pl.tul.emailsemotions.formalityservice.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.tul.emailsemotions.formalityservice.model.Text;
 
-public interface TextRepository extends JpaRepository<Text, Integer> {
+public interface TextRepository extends JpaRepository<Text, Long> {
 }
